AiidaDFTK

Julia-side implementation of the Aiida plugin of DFTK. For more information how to setup and use DFTK with Aiida, see the documentation of the python side of the plugin.

Useful information might also be available in the DFTK-specific guides for Installation and Using DFTK on compute clusters.

Functionality

This package implements a run function, which parses a JSON input to trigger a DFTK-based calculation. Results are again dumped in a JSON/HDF5-compatible form and can be easily parsing in downstream scripts.

At the moment there are no guarantees in the input / output format except that it works in combination with the aiida-dftk plugin. If you wish to use the pipeline implemented by this package in combination with a different downstream driver, please feel free to contact us.

Some general (and not necessarily up to date) remarks on the input and output formats are given in Input and Output format.

API and supported methods

AiidaDFTK.run — Method

run() -> @NamedTuple{output_files::Vector{String}}

Run a DFTK calculation from a json input file. The input file name is expected to be passed as the first argument when calling Julia (i.e. it should be available via ARGS. This function is expected to be called from queuing system jobscripts, for example:

julia --project -e 'using AiidaDFTK; AiidaDFTK.run()' /path/to/input/file.json

It automatically dumps a logfile file.log (i.e. basename of the input file with the log extension), which contains the log messages (i.e. @info, @warn, ...). Currently stdout and stderr are still printed.

AiidaDFTK.run_json — Method

run_json(
    filename::AbstractString;
    extra_output_files
) -> @NamedTuple{output_files::Vector{String}}

Run a DFTK calculation from a json input file. Output is by default written to stdout and stderr. The list of generated output files is returned.