AiidaDFTK

Julia-side implementation of the Aiida plugin of DFTK.

Enables running a DFTK calculation from a JSON file, such that all computed results are again available in JSON or HDF5 for programmatic parsing.

TODO

We should document the expected Julia setup (with regards to MPI, MKL (Blas / FFT), etc.).

API and supported methods

AiidaDFTK.run — Method

run() -> NamedTuple{(:output_files,), Tuple{Vector{String}}}

Run a DFTK calculation from a json input file. The input file name is expected to be passed as the first argument when calling Julia (i.e. it should be available via ARGS. This function is expected to be called from queuing system jobscripts, for example:

julia --project -e 'using AiidaDFTK; AiidaDFTK.run()' /path/to/input/file.json

It automatically dumps a logfile file.log (i.e. basename of the input file with the log extension), which contains the log messages (i.e. @info, @warn, ...). Currently stdout and stderr is still printed.

AiidaDFTK.run_json — Method

run_json(
    filename::AbstractString;
    extra_output_files
) -> NamedTuple{(:output_files,), Tuple{Vector{String}}}

Run a DFTK calculation from a json input file. Output is by default written to stdout and stderr. The list of generated output files is returned.